Moves URL injection to data format helpers
Relocates the `injectAnchorElements` function from the custom field renderers to the data format helpers.

This promotes reusability and avoids code duplication,
as this functionality may be needed in other parts of the application.
